Fixes to regenerate fig/results for reviewers
Removed a duplicate step in the plots which was causing coverage to be set to a single value.

Also modified the package code so that the 'state' variable is converted to a numeric, along with ensuring the unit and time variables are carried through the premodel step correctly. Also, modified how iterations were being set so that the same process is used for both parallel and sequential dispatch methods.